homework_gamer/requirements.txt
attrs==21.2.0
Automat==20.2.0
cffi==1.14.5
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
lxml==4.6.3
parsel==1.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
queuelib==1.6.1
redis==3.5.3
Scrapy==1.6.0
service-identity==18.1.0
six==1.16.0
Twisted==21.2.0
w3lib==1.22.0
zope.interface==5.4.0
